pinned queue, video doesn't loop when disconnected

+++ front/src/lib/playlist/PlaylistItem.svelte
@@ -8,24 +8,33 @@
 	import { user } from '$lib/stores/user';
 
 	export let item: any;
-	export let deleteItem: Function;
+	export let deleteItem: Function | null = null;
+	export let compact = false;
 	let innerWidth = 0;
-	const getURL = (type:string, url:string)=>{
-		if(type =='iframe')
-			return '';
+	const getURL = (type: string, url: string) => {
+		if (type == 'iframe') return '';
 		let provider = '';
-		if(type == 'yt' || type=='ytlive')
-			provider = 'https://youtube.com/watch?v=';
-		if(type == 'tw_live')
-			provider = 'https://twitch.tv/';
+		if (type == 'yt' || type == 'ytlive') provider = 'https://youtube.com/watch?v=';
+		if (type == 'tw_live') provider = 'https://twitch.tv/';
 		return provider + url;
-	}
+	};
 </script>
 
 <svelte:window bind:innerWidth />
-{#if innerWidth < 768}
+{#if compact}
+	<td style={'line-height:normal;padding:0 0.5rem'}>
+		{#if getURL(item.type, item.url)}
+			<a href={getURL(item.type, item.url)} target="_blank" rel="noreferrer">{item.name}</a>
+		{:else}
+			{item.name}
+		{/if}
+		<br />
+		{secondsToTime(item.duration)} - {item.username}
+		<span style:float="right">{new Date(item.startDate).toLocaleTimeString()}</span>
+	</td>
+{:else if innerWidth < 768}
 	<td>
-		{#if $user.accessLevel >= $permissions.manageQueue || $user.username == item.username}
+		{#if ($user.accessLevel >= $permissions.manageQueue || $user.username == item.username) && deleteItem !== null}
 			<!-- svelte-ignore a11y-click-events-have-key-events -->
 			<div class="svgIcon" on:click={deleteItem(item)}>
 				<MdDelete />
@@ -33,22 +42,18 @@
 		{/if}
 	</td>
 	<td colspan="4">
-		{#if getURL(item.type,item.url)}
-			<a
-				href={getURL(item.type,item.url)}
-				target="_blank"
-				rel="noreferrer">{item.name}</a
-			>
+		{#if getURL(item.type, item.url)}
+			<a href={getURL(item.type, item.url)} target="_blank" rel="noreferrer">{item.name}</a>
 		{:else}
 			{item.name}
 		{/if}
 		<br />
 		{secondsToTime(item.duration)} - {item.username}
-		<span style:float='right'>{new Date(item.endDate).toLocaleTimeString()}</span>
+		<span style:float="right">{new Date(item.endDate).toLocaleTimeString()}</span>
 	</td>
 {:else}
-	<td class='t-mid'>
-		{#if $user.accessLevel >= $permissions.manageQueue || $user.username == item.username}
+	<td class="t-mid">
+		{#if ($user.accessLevel >= $permissions.manageQueue || $user.username == item.username) && deleteItem !== null}
 			<!-- svelte-ignore a11y-click-events-have-key-events -->
 			<div class="svgIcon" on:click={deleteItem(item)}>
 				<MdDelete />
@@ -56,14 +61,11 @@
 		{/if}
 	</td>
 	<td>
-		{#if getURL(item.type,item.url)}
-		<a href={getURL(item.type,item.url)}
-			target="_blank"
-			rel="noreferrer">{item.name}
-		</a>
-	{:else}
-		{item.name}
-	{/if}</td
+		{#if getURL(item.type, item.url)}
+			<a href={getURL(item.type, item.url)} target="_blank" rel="noreferrer">{item.name} </a>
+		{:else}
+			{item.name}
+		{/if}</td
 	>
 	<td class="t-right">{new Date(item.startDate).toLocaleTimeString()}</td>
 	<td class="t-right">{new Date(item.endDate).toLocaleTimeString()}</td>
@@ -76,13 +78,13 @@
 		line-height: 2rem;
 		color: var(--color-text-dark-1);
 	}
-	.t-mid{
+	.t-mid {
 		text-align: center;
 	}
 	.t-right {
 		text-align: right;
 	}
-	.svgIcon{
-		display:inline-block;
+	.svgIcon {
+		display: inline-block;
 	}
 </style>
